requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.67
langchain-openai==0.3.27
langgraph==0.2.39
langgraph-checkpoint==2.1.0
langgraph-sdk==0.1.72
langsmith==0.4.4
openai==1.93.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.38.6
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.32.0
zstandard==0.23.0
